Shifted common functions to newly created common module
