+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/common/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@egovernments/digit-ui-module-common",
+  "version": "0.0.1",
+  "description": "Common Utilities Module UI",
+  "main": "dist/index.js",
+  "module": "dist/index.modern.js",
+  "source": "src/Module.js",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "start": "microbundle-crl watch --no-compress --format modern",
+    "build": "microbundle-crl --compress --no-sourcemap --format cjs",
+    "prepublish": "yarn build"
+  },
+  "peerDependencies": {
+    "react": "17.0.2",
+    "react-router-dom": "5.3.0"
+  },
+  "dependencies": {
+    "@egovernments/digit-ui-react-components": "1.8.2-beta.11",
+    "@egovernments/digit-ui-components": "0.0.2-beta.1",
+    "@egovernments/digit-ui-module-core": "1.8.1-beta.6",
+    "react": "17.0.2",
+    "react-dom": "17.0.2"
+  },
+  "author": "eGov Foundation",
+  "license": "MIT"
+}
