+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/common/src/utils/index.js
@@ -1,0 +1,13 @@
+export const getFormattedName = (firstName, middleName, lastName, designation, partyTypeLabel) => {
+  const nameParts = [firstName, middleName, lastName]
+    ?.map((part) => part?.trim())
+    ?.filter(Boolean)
+    ?.join(" ")
+    ?.trim();
+
+  const nameWithDesignation = designation && nameParts ? `${nameParts} - ${designation}` : designation || nameParts;
+
+  return partyTypeLabel ? `${nameWithDesignation} ${partyTypeLabel}` : nameWithDesignation;
+};
+
+export { getAllAssignees, getAdvocates } from "./caseUtils";
