+++ frontend/micro-ui/web/package.json
@@ -18,6 +18,7 @@
     "@egovernments/digit-ui-components": "0.0.1-beta.28",
     "@egovernments/digit-ui-libraries": "1.8.1-beta.3",
     "@egovernments/digit-ui-module-core": "1.8.1-beta.6",
+    "@egovernments/digit-ui-module-common": "0.0.1",
     "@egovernments/digit-ui-module-dristi": "0.0.1",
     "@egovernments/digit-ui-module-hearings": "0.0.1",
     "@egovernments/digit-ui-module-home": "0.0.1",
